requirements.txt
aiosqlite==0.17.0
anyio==3.3.2
asgiref==3.4.1
asyncmy==0.2.1
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
fastapi-crudrouter==0.8.4
h11==0.12.0
idna==3.2
iso8601==0.1.16
pydantic==1.8.2
pypika-tortoise==0.1.1
python-dotenv==0.19.1
pytz==2021.3
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
